+++ apps/backoffice-v2/src/common/components/molecules/Pagination/Pagination.tsx
@@ -10,4 +10,5 @@ export const Pagination = ({ className, ...props }: ComponentProps<'nav'>) => (
     {...props}
   />
 );
+
 Pagination.displayName = 'Pagination';