+++ apps/backoffice-v2/src/lib/blocks/variants/DefaultBlocks/hooks/useDefaultBlocksLogic/useDefaultBlocksLogic.tsx
@@ -48,6 +48,8 @@ import { useCaseOverviewBlock } from '@/lib/blocks/hooks/useCaseOverviewBlock/us
 import { useSearchParamsByEntity } from '@/common/hooks/useSearchParamsByEntity/useSearchParamsByEntity';
 import { useLocation } from 'react-router-dom';
 import { omitPropsFromObjectWhitelist } from '@/common/utils/omit-props-from-object-whitelist/omit-props-from-object-whitelist';
+import { useObjectEntriesBlock } from '@/lib/blocks/hooks/useObjectEntriesBlock/useObjectEntriesBlock';
+import { useAmlBlock } from '@/lib/blocks/components/AmlBlock/hooks/useAmlBlock/useAmlBlock';
 
 const registryInfoWhitelist = ['open_corporates'] as const;
 
@@ -391,6 +393,29 @@ export const useDefaultBlocksLogic = () => {
 
   const caseOverviewBlock = useCaseOverviewBlock();
 
+  const customDataBlock = useObjectEntriesBlock({
+    object: workflow?.context?.customData ?? {},
+    heading: 'Custom Data',
+  });
+
+  const amlData = useMemo(() => [workflow?.context?.aml], [workflow?.context?.aml]);
+
+  const amlBlock = useAmlBlock(amlData);
+
+  const amlWithContainerBlock = useMemo(() => {
+    if (!amlBlock?.length) {
+      return [];
+    }
+
+    return createBlocksTyped()
+      .addBlock()
+      .addCell({
+        type: 'block',
+        value: amlBlock,
+      })
+      .build();
+  }, [amlBlock]);
+
   const allBlocks = useMemo(() => {
     if (!workflow?.context?.entity) return [];
 
@@ -420,6 +445,8 @@ export const useDefaultBlocksLogic = () => {
       documentReviewBlocks,
       businessInformationBlocks,
       caseOverviewBlock,
+      customDataBlock,
+      amlWithContainerBlock,
     ];
   }, [
     associatedCompaniesBlock,
@@ -447,6 +474,8 @@ export const useDefaultBlocksLogic = () => {
     documentReviewBlocks,
     businessInformationBlocks,
     caseOverviewBlock,
+    customDataBlock,
+    amlWithContainerBlock,
     workflow?.context?.entity,
   ]);