+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ballerine/headless-example
 
+## 0.3.24
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.16
+  - @ballerine/workflow-browser-sdk@0.6.25
+
+## 0.3.23
+
+### Patch Changes
+
+- @ballerine/workflow-browser-sdk@0.6.24
+
+## 0.3.22
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.15
+  - @ballerine/workflow-browser-sdk@0.6.23
+
 ## 0.3.21
 
 ### Patch Changes