+++ packages/common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/common
 
+## 0.9.16
+
+### Patch Changes
+
+- Bump
+
+## 0.9.15
+
+### Patch Changes
+
+- Bump
+
 ## 0.9.14
 
 ### Patch Changes