+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/eslint-config
 
+## 1.1.9
+
+### Patch Changes
+
+- Bump
+
+## 1.1.8
+
+### Patch Changes
+
+- Bump
+
 ## 1.1.7
 
 ### Patch Changes