+++ packages/rules-engine/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/rules-engine-lib
 
+## 0.5.9
+
+### Patch Changes
+
+- Bump
+
+## 0.5.8
+
+### Patch Changes
+
+- Bump
+
 ## 0.5.7
 
 ### Patch Changes