+++ services/websocket-service/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/websocket-service
 
+## 0.1.9
+
+### Patch Changes
+
+- Bump
+
+## 0.1.8
+
+### Patch Changes
+
+- Bump
+
 ## 0.1.7
 
 ### Patch Changes